kit list working
