+++ lib/apolloClient.js
@@ -14,6 +14,7 @@ function createApolloClient() {
     link: new HttpLink({
       uri: 'https://rocket-club-online.hasura.app/v1/graphql', // Server URL (must be absolute)
       credentials: 'same-origin', // Additional fetch() options like `credentials` or `headers`
+      headers: {'x-hasura-admin-secret': 'ePHT1XS24pOVmXrgTMXV6qxtREujN257tWekhWXgeA1ut8D1WWM6Lhc7xzA71Oke'}
     }),
     cache: new InMemoryCache({
       typePolicies: {
