--- lib/withApollo.js
@@ -1,184 +1,0 @@
-import React from "react";
-import App from "next/app";
-import Head from "next/head";
-import { ApolloProvider } from "@apollo/react-hooks";
-import createApolloClient from "./apolloClient";
-import { getSession } from '@auth0/nextjs-auth0';
-
-// On the client, we store the Apollo Client in the following variable.
-// This prevents the client from reinitializing between page transitions.
-let globalApolloClient = null;
-
-/**
- * Installs the Apollo Client on NextPageContext
- * or NextAppContext. Useful if you want to use apolloClient
- * inside getStaticProps, getStaticPaths or getServerSideProps
- * @param {NextPageContext | NextAppContext} ctx
- */
-export const initOnContext = (ctx) => {
-  const inAppContext = Boolean(ctx.ctx);
-
-  // We consider installing `withApollo({ ssr: true })` on global App level
-  // as antipattern since it disables project wide Automatic Static Optimization.
-  if (process.env.NODE_ENV === "development") {
-    if (inAppContext) {
-      console.warn(
-        "Warning: You have opted-out of Automatic Static Optimization due to `withApollo` in `pages/_app`.\n" +
-          "Read more: https://err.sh/next.js/opt-out-auto-static-optimization\n"
-      );
-    }
-  }
-
-  // Initialize ApolloClient if not already done
-  const apolloClient =
-    ctx.apolloClient ||
-    initApolloClient(ctx.apolloState || {}, inAppContext ? ctx.ctx : ctx);
-
-  // We send the Apollo Client as a prop to the component to avoid calling initApollo() twice in the server.
-  // Otherwise, the component would have to call initApollo() again but this
-  // time without the context. Once that happens, the following code will make sure we send
-  // the prop as `null` to the browser.
-  apolloClient.toJSON = () => null;
-
-  // Add apolloClient to NextPageContext & NextAppContext.
-  // This allows us to consume the apolloClient inside our
-  // custom `getInitialProps({ apolloClient })`.
-  ctx.apolloClient = apolloClient;
-  if (inAppContext) {
-    ctx.ctx.apolloClient = apolloClient;
-  }
-
-  return ctx;
-};
-
-async function getHeaders(ctx) {
-  if (typeof window !== "undefined") return null;
-  if (typeof ctx.req === "undefined") return null;
-
-  const s = getSession(ctx.req, ctx.res);
-  if (s && s.accessToken == null) return null;
-
-  return {
-    authorization: `Bearer ${s ? s.accessToken : ""}`,
-  };
-}
-
-/**
- * Always creates a new apollo client on the server
- * Creates or reuses apollo client in the browser.
- * @param  {NormalizedCacheObject} initialState
- * @param  {NextPageContext} ctx
- */
-const initApolloClient = (initialState, headers) => {
-  // Make sure to create a new client for every server-side request so that data
-  // isn't shared between connections (which would be bad)
-  if (typeof window === "undefined") {
-    return createApolloClient(initialState, headers);
-  }
-
-  // Reuse client on the client-side
-  if (!globalApolloClient) {
-    globalApolloClient = createApolloClient(initialState, headers);
-  }
-
-  return globalApolloClient;
-};
-
-/**
- * Creates a withApollo HOC
- * that provides the apolloContext
- * to a next.js Page or AppTree.
- * @param  {Object} withApolloOptions
- * @param  {Boolean} [withApolloOptions.ssr=false]
- * @returns {(PageComponent: ReactNode) => ReactNode}
- */
-export const withApollo =
-  ({ ssr = true } = {}) =>
-  (PageComponent) => {
-    const WithApollo = ({ apolloClient, apolloState, ...pageProps }) => {
-      let client;
-      if (apolloClient) {
-        // Happens on: getDataFromTree & next.js ssr
-        client = apolloClient;
-      } else {
-        // Happens on: next.js csr
-        // client = initApolloClient(apolloState, undefined);
-        client = initApolloClient(apolloState, {});
-      }
-
-      return (
-        <ApolloProvider client={client}>
-          <PageComponent {...pageProps} />
-        </ApolloProvider>
-      );
-    };
-
-    // Set the correct displayName in development
-    if (process.env.NODE_ENV !== "production") {
-      const displayName =
-        PageComponent.displayName || PageComponent.name || "Component";
-      WithApollo.displayName = `withApollo(${displayName})`;
-    }
-    if (ssr || PageComponent.getInitialProps) {
-      WithApollo.getInitialProps = async (ctx) => {
-        const { AppTree } = ctx;
-
-        // Initialize ApolloClient, add it to the ctx object so
-        // we can use it in `PageComponent.getInitialProp`.
-        const apolloClient = (ctx.apolloClient = initApolloClient(
-          null,
-          await getHeaders(ctx)
-        ));
-
-        // Run wrapped getInitialProps methods
-        let pageProps = {};
-        if (PageComponent.getInitialProps) {
-          pageProps = await PageComponent.getInitialProps(ctx);
-        }
-
-        // Only on the server:
-        if (typeof window === "undefined") {
-          // When redirecting, the response is finished.
-          // No point in continuing to render
-          if (ctx.res && ctx.res.finished) {
-            return pageProps;
-          }
-
-          // Only if ssr is enabled
-          if (ssr) {
-            try {
-              // Run all GraphQL queries
-              const { getDataFromTree } = await import("@apollo/react-ssr");
-              await getDataFromTree(
-                <AppTree
-                  pageProps={{
-                    ...pageProps,
-                    apolloClient,
-                  }}
-                />
-              );
-            } catch (error) {
-              // Prevent Apollo Client GraphQL errors from crashing SSR.
-              // Handle them in components via the data.error prop:
-              // https://www.apollographql.com/docs/react/api/react-apollo.html#graphql-query-data-error
-              console.error("Error while running `getDataFromTree`", error);
-            }
-
-            // getDataFromTree does not call componentWillUnmount
-            // head side effect therefore need to be cleared manually
-            Head.rewind();
-          }
-        }
-
-        // Extract query data from the Apollo store
-        const apolloState = apolloClient.cache.extract();
-
-        return {
-          ...pageProps,
-          apolloState,
-        };
-      };
-    }
-
-    return WithApollo;
-  };
