+++ pages/index.js
@@ -1,17 +1,29 @@
-import Link from "next/link";
-import Date from "../components/date";
-import Head from "next/head";
 import Layout, { siteTitle } from "../components/layout";
-import { initializeApollo, addApolloState } from '../lib/apolloClient'
+import { initializeApollo, addApolloState } from "../lib/apolloClient";
+import KitList, {
+  ALL_KITS_QUERY,
+} from "../components/KitList";
 
 const Home = () => {
   return (
     <Layout>
       <div className="bg-white overflow-hidden shadow rounded-lg">
- 
       </div>
     </Layout>
   );
+};
+
+export async function getStaticProps() {
+  const apolloClient = initializeApollo();
+
+  await apolloClient.query({
+    query: ALL_KITS_QUERY,
+  });
+
+  return addApolloState(apolloClient, {
+    props: {},
+    revalidate: 1,
+  });
 }
 
-export default withApollo({ ssr: true })(Home)
+export default Home;
